Rewrites tracer tests to assert in v2
